requirements.txt
streamlit
fpdf2>=2.7.0
